Separar menu en componentes
