+++ src/views/AppAppBar.tsx
@@ -1,16 +1,12 @@
-import MenuIcon from "@mui/icons-material/Menu";
-import {
-  Container,
-  IconButton,
-  Menu,
-  MenuItem,
-  Typography,
-} from "@mui/material";
+import { Container, Link } from "@mui/material";
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import React from "react";
 import AppBar from "../components/AppBar";
 import Toolbar from "../components/Toolbar";
+import { useRouter } from 'next/router';
+import MenuHamburguesa from "@/components/MenuHamburguesa";
+import Menu from "@/components/Menu";
 
 const pages = [
   {
@@ -31,18 +27,27 @@ const pages = [
   }
 ];
 
-function AppAppBar() {
-  const [anchorElNav, setAnchorElNav] = React.useState<null | HTMLElement>(
-    null
-  );
+type AppAppBarProps = {
+  mostrarMenu?: boolean;
+  mostrarMenuHamburguesa?: boolean;
+}
 
-  const handleOpenNavMenu = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorElNav(event.currentTarget);
-  };
+function AppAppBar(props: AppAppBarProps) {
 
-  const handleCloseNavMenu = () => {
-    setAnchorElNav(null);
-  };
+  const {
+    mostrarMenu = true,
+    mostrarMenuHamburguesa = true
+  } = props;
+
+  const router = useRouter();
+
+  const handleSignInClick = () => {
+    router.push('/SignIn');
+  }
+
+  const handleRegisterClick = () => {
+    router.push('/Register')
+  }
 
   return (
     <>
@@ -55,7 +60,8 @@ function AppAppBar() {
             padding: "11px 0",
           }}
         >
-          <Box sx={{flex: 1, display: "flex" }}>
+
+          <Link href="/" sx={{ flex: 1, display: "flex" }}>
             <Box
               component="img"
               src="/logo1.svg"
@@ -66,117 +72,25 @@ function AppAppBar() {
               component="img"
               src="/logo2.svg"
               alt="nombre"
-              sx={{ width: "150px", marginLeft: "20px", display:{ xs: "none", md: "block"} }}
+              sx={{ width: "150px", marginLeft: "20px", display: { xs: "none", md: "block" } }}
             />
-          </Box>
+          </Link>
+
+
           <Box sx={{ justifyContent: "flex-end", alignItems: "center" }}>
-            <Button sx={{ color: "#fff", fontSize: "1.3rem" }}>Ingresar</Button>
-            <Button sx={{ color: "secondary.main", fontSize: "1.3rem" }}>
+            <Button sx={{ color: "#fff", fontSize: "1.3rem" }} onClick={handleSignInClick}>Ingresar</Button>
+            <Button sx={{ color: "secondary.main", fontSize: "1.3rem" }} onClick={handleRegisterClick}>
               Registrarse
             </Button>
           </Box>
-          {/* Menu Hamburguesa */}
-          <Box sx={{ display: { xs: "flex", md: "none" } }}>
-            <IconButton
-              size="large"
-              aria-label="Menu"
-              aria-controls="menu-appbar"
-              aria-haspopup="true"
-              onClick={handleOpenNavMenu}
-              color="inherit"
-            >
-              <MenuIcon />
-            </IconButton>
-            <Menu
-              id="menu-appbar"
-              slotProps={{
-                paper: {
-                  sx: {
-                    width: "100vw",
-                    maxWidth: "100vw",
-                    left: "0 !important",
-                    top: 0,
-                    marginTop: "19px",
-                    position: "fixed",
-                    borderRadius: 0,
-                    background: "rgba(0, 53, 102, 0.8)",
-                    color: "#FFF",
-                  },
-                },
-              }}
-              anchorEl={anchorElNav}
-              anchorOrigin={{
-                vertical: "bottom",
-                horizontal: "left",
-              }}
-              keepMounted
-              transformOrigin={{
-                vertical: "top",
-                horizontal: "left",
-              }}
-              open={Boolean(anchorElNav)}
-              onClose={handleCloseNavMenu}
-              sx={{
-                display: { xs: "block", md: "none" },
-              }}
-            >
-              {pages.map((page) => (
-                <MenuItem
-                  key={page.id}
-                  onClick={handleCloseNavMenu}
-                  sx={{ justifyContent: "flex-end", marginRight: "18px" }}
-                >
-                  <Typography textAlign="center">
-                    <a href={`#${page.id}`}>{page.title}</a> 
-                  </Typography>
-                </MenuItem>
-              ))}
-            </Menu>
-          </Box>
+
+          {mostrarMenuHamburguesa &&
+            <MenuHamburguesa pages={pages} />}
         </Toolbar>
 
-        {/* Menu */}
-        <Box
-          sx={{
-            flexGrow: 1,
-            display: { xs: "none", md: "flex" },
-            justifyContent: "center",
-            background: "rgba(0, 53, 102, 0.47)",
-            padding: "3px 0",
-          }}
-        >
-          <Container>
-            <ul
-              style={{
-                listStyleType: "none",
-                padding: 0,
-                margin: 0,
-                display: "flex",
-                justifyContent: "center",
-              }}
-            >
-              {pages.map((page) => (
-                <li key={page.id}>
-                  <a
-                    href={`#${page.id}`}
-                    onClick={handleCloseNavMenu}
-                    style={{
-                      margin: "2px 32px",
-                      color: "white",
-                      display: "block",
-                      fontSize: "1.2rem",
-                      textTransform: "none",
-                      fontWeight: 600,
-                      textDecoration: "none",
-                    }}
-                  >
-                    {page.title}
-                  </a>
-                </li>
-              ))}
-            </ul>
-          </Container>
-        </Box>
+        {mostrarMenu &&
+         <Menu pages={pages} />}
+
       </AppBar>
     </>
   );
